+++ src/test/kotlin/org/kalasim/examples/api/CustomEvent.kts
@@ -1,9 +1,18 @@
-import org.kalasim.Event
-import org.kalasim.TickTime
-import org.kalasim.createSimulation
+import org.kalasim.*
 
 class MyEvent(val context: String, time: TickTime) : Event(time)
 
 createSimulation {
+
+    object : Component("Something") {
+        override fun process() = sequence<Component> {
+            //... a great history
+            log(MyEvent("foo", now))
+            //... a promising future
+        }
+    }
+
     addEventListener<MyEvent> { println(it.context) }
+
+    run()
 }